refactor: Move Mapbox components to special folder
refactor: Move Mapbox components into dedicated directory

Relocates the primary map components from the root 'app/' directory to 'app/TestMaps/' for better organization and separation of concerns.

Files moved:
- app/MyMapbox.tsx
- app/MyMapboxGeofence.tsx
- app/MyMapboxPopup.tsx
