+++ app/page.tsx
@@ -1,10 +1,10 @@
 import "./pagestyle.css";
 
-import { MyMapComponent } from "./MyMapbox";
+import { MyMapComponent } from "./TestMaps/MyMapbox";
 
-import { MyMapboxPopup } from "./MyMapboxPopup";
+import { MyMapboxPopup } from "./TestMaps/MyMapboxPopup";
 
-import { MyMapboxGeofence } from "./MyMapboxGeofence";
+import { MyMapboxGeofence } from "./TestMaps/MyMapboxGeofence";
 
 const MyApp = function () {
   // return <div className="header"> </div>;
@@ -27,8 +27,8 @@ const MyApp = function () {
           <div className="content">
             <h2>Main Content</h2>
             {/* <MyMapComponent /> */}
-            {/* <MyMapboxPopup /> */}
-            <MyMapboxGeofence />
+            <MyMapboxPopup />
+            {/* <MyMapboxGeofence /> */}
           </div>
         </div>
         <div className="footer">
